Implemented write syscall

+++ src/systemcall/syscall_no.rs
@@ -1,7 +1,7 @@
 use num_enum::TryFromPrimitive;
 
 #[derive(Debug, Clone, Copy, PartialEq, TryFromPrimitive)]
-#[repr(isize)]
+#[repr(usize)]
 pub enum SystemCallNo {
-    Write = 1 as isize,
+    Write = 1,
 }
